Move pytest fixture setup_fall_detection_lib_path
